shadowlink-ai/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "shadowlink-ai"
version = "0.1.0"
description = "ShadowLink AI Service — Agent orchestration, RAG pipeline, MCP protocol, and tool execution"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "ShadowLink Team" }]
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # ── Web Framework ──
    "fastapi>=0.111,<1",
    "uvicorn[standard]>=0.29,<1",
    "sse-starlette>=2,<3",
    "python-multipart>=0.0.9",

    # ── Data Validation & Settings ──
    "pydantic>=2.7,<3",
    "pydantic-settings>=2.2,<3",

    # ── LLM & Agent Framework ──
    "langchain>=0.2,<1",
    "langchain-core>=0.2,<1",
    "langchain-openai>=0.1,<1",
    "langchain-community>=0.2,<1",
    "langgraph>=0.2,<1",
    "langgraph-checkpoint>=1,<3",

    # ── gRPC ──
    "grpcio>=1.63,<2",
    "grpcio-tools>=1.63,<2",
    "grpcio-reflection>=1.63,<2",
    "protobuf>=5.26,<7",

    # ── Vector Search & Embeddings ──
    "faiss-cpu>=1.8,<2",
    "sentence-transformers>=3,<4",

    # ── HTTP Client ──
    "httpx>=0.27,<1",

    # ── File Processing ──
    "pymupdf>=1.24,<2",
    "python-docx>=1.1,<2",
    "openpyxl>=3.1,<4",
    "python-pptx>=0.6,<1",
    "markdown>=3.6,<4",

    # ── Structured Logging ──
    "structlog>=24,<26",

    # ── MCP Protocol ──
    "mcp>=1.0,<2",

    # ── Knowledge Graph ──
    "networkx>=3.2,<4",

    # ── Utilities ──
    "tenacity>=8.3,<10",
    "orjson>=3.10,<4",
    "aiofiles>=23,<25",
    "cachetools>=5.3,<6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
    "pytest-cov>=5,<6",
    "pytest-mock>=3.14,<4",
    "httpx>=0.27,<1",
    "ruff>=0.4,<1",
    "mypy>=1.10,<2",
    "pre-commit>=3.7,<4",
]

rerank = [
    # ── Reranking (optional, heavy) ──
    "FlagEmbedding>=1.2,<2",
]

ocr = [
    # ── OCR (optional, heavy) ──
    "paddleocr>=2.7,<3",
    "paddlepaddle>=2.6,<3",
]

all = [
    "shadowlink-ai[dev,rerank,ocr]",
]

[project.scripts]
shadowlink-ai = "app.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# ── Ruff (Linter + Formatter) ──
[tool.ruff]
target-version = "py310"
line-length = 120
src = ["app", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "T20", # flake8-print
    "RUF", # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument (FastAPI Depends)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ── MyPy ──
[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "grpc.*",
    "google.protobuf.*",
    "faiss.*",
    "sentence_transformers.*",
    "langchain.*",
    "langchain_core.*",
    "langchain_openai.*",
    "langchain_community.*",
    "langgraph.*",
    "mcp.*",
    "pymupdf.*",
    "fitz.*",
    "docx.*",
    "openpyxl.*",
    "pptx.*",
    "FlagEmbedding.*",
]
ignore_missing_imports = true

# ── Pytest ──
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "unit: Unit tests (fast, no external deps)",
    "integration: Integration tests (may need services)",
    "e2e: End-to-end tests (full stack)",
    "slow: Slow tests (model loading, etc.)",
]
filterwarnings = [
    "ignore::DeprecationWarning:langchain.*",
]

# ── Coverage ──
[tool.coverage.run]
source = ["app"]
omit = ["app/api/grpc/*_pb2*", "tests/*"]

[tool.coverage.report]
fail_under = 60
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
    "pass",
    "\\.\\.\\.",
]
